added drawer which is independent from main content
