+++ client/src/components/Nav/NavBar.jsx
@@ -20,50 +20,7 @@ import ProfileDialog from "../Dashboard/ProfileDialog";
 import logo from "../../assets/logo.jpg";
 import socket from "../../utils/socket";
 
-const drawerWidth = 240;
-
-const useStyles = makeStyles(theme => ({
-  root: {
-    display: "flex"
-  },
-  appBar: {
-    zIndex: theme.zIndex.drawer + 1,
-    backgroundColor: "white",
-    position: "relative"
-  },
-  logo: {
-    height: "4rem",
-    width: "4rem",
-    marginLeft: "5rem",
-    position: "absolute"
-  },
-  menuButton: {
-    marginRight: theme.spacing(2),
-    textDecoration: "none"
-  },
-  drawer: {
-    width: drawerWidth,
-    flexShrink: 0
-  },
-  drawerPaper: {
-    width: drawerWidth
-  },
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3)
-  },
-  toolbar: {
-    marginLeft: "auto",
-    marginRight: "5rem",
-    ...theme.mixins.toolbar
-  },
-  name: {
-    marginLeft: "0.5rem"
-  }
-}));
-
-const NavBar = ({ user, handleLogOut, match }) => {
-  const classes = useStyles();
+const NavBar = ({ classes, user, handleLogOut, match }) => {
   const [open, setOpen] = useState(false);
   const anchorRef = useRef(null);
 
@@ -94,69 +51,67 @@ const NavBar = ({ user, handleLogOut, match }) => {
   };
 
   return (
-    <div className={classes.root}>
-      <AppBar position="static" className={classes.appBar}>
-        <NavLink exact to="/dashboard">
-          <img src={logo} alt="logo" className={classes.logo} />
+    <AppBar position="fixed" className={classes.appBar}>
+      <NavLink exact to="/dashboard">
+        <img src={logo} alt="logo" className={classes.logo} />
+      </NavLink>
+      <Toolbar className={classes.toolbar}>
+        <NavLink exact to="/friends" className={classes.menuButton}>
+          <Button>Friends</Button>
         </NavLink>
-        <Toolbar className={classes.toolbar}>
-          <NavLink exact to="/friends" className={classes.menuButton}>
-            <Button>Friends</Button>
-          </NavLink>
 
-          <NavLink exact to="/friends-polls" className={classes.menuButton}>
-            <Button>Friends poll</Button>
-          </NavLink>
+        <NavLink exact to="/friends-polls" className={classes.menuButton}>
+          <Button>Friends poll</Button>
+        </NavLink>
 
-          <ProfileDialog user={user} className={classes.menuButton} />
+        <ProfileDialog user={user} className={classes.menuButton} />
 
-          <Button
-            className={classes.name}
-            ref={anchorRef}
-            aria-controls={open ? "menu-list-grow" : undefined}
-            aria-haspopup="true"
-            onClick={handleToggle}
-          >
-            <OnlineBadge overlap="circle" variant="dot">
-              <Avatar alt={user.name} src={user.avatar} />
-            </OnlineBadge>
-            <Typography className={classes.name}>{user.name}</Typography>
-          </Button>
+        <Button
+          className={classes.name}
+          ref={anchorRef}
+          aria-controls={open ? "menu-list-grow" : undefined}
+          aria-haspopup="true"
+          onClick={handleToggle}
+        >
+          <OnlineBadge overlap="circle" variant="dot">
+            <Avatar alt={user.name} src={user.avatar} />
+          </OnlineBadge>
+          <Typography className={classes.name}>{user.name}</Typography>
+        </Button>
 
-          <Popper
-            open={open}
-            anchorEl={anchorRef.current}
-            role={undefined}
-            transition
-            disablePortal
-          >
-            {({ TransitionProps, placement }) => (
-              <Grow
-                {...TransitionProps}
-                style={{
-                  transformOrigin:
-                    placement === "bottom" ? "center top" : "center bottom"
-                }}
-              >
-                <Paper>
-                  <ClickAwayListener onClickAway={handleClose}>
-                    <MenuList id="menu-list-grow">
-                      <MenuItem
-                        onClick={() => {
-                          handleLogOut();
-                        }}
-                      >
-                        Logout
-                      </MenuItem>
-                    </MenuList>
-                  </ClickAwayListener>
-                </Paper>
-              </Grow>
-            )}
-          </Popper>
-        </Toolbar>
-      </AppBar>
-    </div>
+        <Popper
+          open={open}
+          anchorEl={anchorRef.current}
+          role={undefined}
+          transition
+          disablePortal
+        >
+          {({ TransitionProps, placement }) => (
+            <Grow
+              {...TransitionProps}
+              style={{
+                transformOrigin:
+                  placement === "bottom" ? "center top" : "center bottom"
+              }}
+            >
+              <Paper>
+                <ClickAwayListener onClickAway={handleClose}>
+                  <MenuList id="menu-list-grow">
+                    <MenuItem
+                      onClick={() => {
+                        handleLogOut();
+                      }}
+                    >
+                      Logout
+                    </MenuItem>
+                  </MenuList>
+                </ClickAwayListener>
+              </Paper>
+            </Grow>
+          )}
+        </Popper>
+      </Toolbar>
+    </AppBar>
   );
 };
