+++ client/src/pages/Dashboard.jsx
@@ -2,15 +2,59 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import { CircularProgress } from "@material-ui/core";
+import { makeStyles } from "@material-ui/core/styles";
 import NavBar from "../components/Nav/NavBar";
-import ContentContainer from "../components/Dashboard/ContentContainer";
 import DashboardDefault from "../components/Dashboard/DashboardDefault";
 import Friends from "./Friends";
 import VotePage from "./VotePage";
 import FriendsPolls from "./FriendsPolls";
+import FriendsDrawer from "../components/FriendsDrawer/FriendsDrawer";
 import socket from "../utils/socket";
 
+const drawerWidth = 200;
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: "flex"
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    backgroundColor: "white"
+  },
+  logo: {
+    height: "4rem",
+    width: "4rem",
+    marginLeft: "5rem",
+    position: "absolute"
+  },
+  menuButton: {
+    marginRight: theme.spacing(2),
+    textDecoration: "none"
+  },
+  content: {
+    flexGrow: 1,
+    padding: theme.spacing(3)
+  },
+  toolbar: {
+    marginLeft: "auto",
+    marginRight: "5rem",
+    ...theme.mixins.toolbar
+  },
+  name: {
+    marginLeft: "0.5rem"
+  },
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0
+  },
+  drawerPaper: {
+    width: drawerWidth
+  }
+}));
+
 const Dashboard = ({ history }) => {
+  const classes = useStyles();
+
   const [user, setUser] = useState(null);
   const [error, setError] = useState(false);
   const [friends, setFriends] = useState([]);
@@ -80,33 +124,40 @@ const Dashboard = ({ history }) => {
     <CircularProgress />
   ) : (
     <Router>
-      <NavBar user={user} handleLogOut={handleLogOut} />
-      <ContentContainer friends={friends} setFriends={setFriends}>
-        <Route
-          exact
-          path="/dashboard"
-          render={() => <DashboardDefault user={user} />}
-        />
-        <Route
-          exact
-          path="/friends"
-          render={props => (
-            <Friends
-              {...props}
-              friends={friends}
-              handleAddorRemoveFriend={handleAddorRemoveFriend}
-            />
-          )}
-        />
+      <div className={classes.root}>
+        <NavBar classes={classes} user={user} handleLogOut={handleLogOut} />
+        <FriendsDrawer classes={classes} friends={friends} />
+        <main className={classes.content}>
+          <Route
+            exact
+            path="/dashboard"
+            render={() => <DashboardDefault user={user} />}
+          />
+          <Route
+            exact
+            path="/friends"
+            render={props => (
+              <Friends
+                {...props}
+                friends={friends}
+                handleAddorRemoveFriend={handleAddorRemoveFriend}
+              />
+            )}
+          />
 
-        <Route exact path={"/friends-polls"} render={() => <FriendsPolls />} />
+          <Route
+            exact
+            path={"/friends-polls"}
+            render={() => <FriendsPolls />}
+          />
 
-        <Route
-          exact
-          path="/polls/:pollId"
-          render={() => <VotePage user={user} />}
-        />
-      </ContentContainer>
+          <Route
+            exact
+            path="/polls/:pollId"
+            render={() => <VotePage user={user} />}
+          />
+        </main>
+      </div>
     </Router>
   );
 };
